fixed VC++ size_t -> int conversion warning
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@30078 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ include/wx/docview.h
@@ -565,7 +565,7 @@ public:
     // Operations
     virtual void AddFileToHistory(const wxString& file);
     virtual void RemoveFileFromHistory(size_t i);
-    virtual int GetMaxFiles() const { return m_fileMaxFiles; }
+    virtual int GetMaxFiles() const { return (int)m_fileMaxFiles; }
     virtual void UseMenu(wxMenu *menu);
 
     // Remove menu from the list (MDI child may be closing)
